Rename: recipe 패키지 관련 패키지를 recipe 패키지에 통합
