+++ server/src/main/java/com/main33/server/recipe/recipe/domain/RecipeFoodType.java
@@ -1,4 +1,4 @@
-package com.main33.server.recipe.domain;
+package com.main33.server.recipe.recipe.domain;
 
 public enum RecipeFoodType {
     KOREAN, JAPANESE, CHINESE, WESTERN, ETC
