+++ server/src/main/java/com/main33/server/recipe/recipe_comment/entity/RecipeComment.java
@@ -1,33 +1,42 @@
-package com.main33.server.recipe_like.entity;
+package com.main33.server.recipe.recipe_comment.entity;
+
 
 import com.main33.server.member.entity.Member;
-import com.main33.server.recipe.entity.Recipe;
+import com.main33.server.recipe.recipe.entity.Recipe;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
 
 import javax.persistence.*;
 import java.sql.Timestamp;
 
 @Entity
-@Table(name = "recipe_like")
+@Table(name = "recipe_comment")
 @Getter
 @Setter
 @NoArgsConstructor
-public class RecipeLike {
+public class RecipeComment {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long likeId;
+    private Long commentId;
 
     @ManyToOne
     @JoinColumn(name="member_id", nullable = false)
     private Member member;
 
     @ManyToOne
-    @JoinColumn(name="recipe_id")
+    @JoinColumn(name="recipe_id", nullable = false)
     private Recipe recipe;
+
+    @Column(nullable = false)
+    private String commentBody;
 
     @CreationTimestamp
     private Timestamp createdAt;
+
+    @UpdateTimestamp
+    private Timestamp modifiedAt;
 }
