+++ server/src/main/java/com/main33/server/recipe/recipe_step/entity/RecipeStep.java
@@ -1,6 +1,7 @@
-package com.main33.server.recipe_image.entity;
+package com.main33.server.recipe.recipe_step.entity;
 
-import com.main33.server.recipe.entity.Recipe;
+
+import com.main33.server.recipe.recipe.entity.Recipe;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -11,21 +12,24 @@ import javax.persistence.*;
 import java.sql.Timestamp;
 
 @Entity
-@Table(name = "recipe_image")
+@Table(name = "recipe_step")
 @Getter
 @Setter
 @NoArgsConstructor
-public class RecipeImage {
+public class RecipeStep {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long imageId;
+    private Long stepId;
 
     @ManyToOne
     @JoinColumn(name="recipe_id", nullable = false)
     private Recipe recipe;
 
-    @Column
-    private String imageUrl;
+    @Column(nullable = false)
+    private Integer stepNumber;
+
+    @Column(nullable = false)
+    private String stepContent;
 
     @CreationTimestamp
     private Timestamp createdAt;
