--- server/src/main/java/com/main33/server/recipe_step/entity/RecipeStep.java
@@ -1,39 +1,0 @@
-package com.main33.server.recipe_step.entity;
-
-
-import com.main33.server.recipe.entity.Recipe;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import javax.persistence.*;
-import java.sql.Timestamp;
-
-@Entity
-@Table(name = "recipe_step")
-@Getter
-@Setter
-@NoArgsConstructor
-public class RecipeStep {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long stepId;
-
-    @ManyToOne
-    @JoinColumn(name="recipe_id", nullable = false)
-    private Recipe recipe;
-
-    @Column(nullable = false)
-    private Integer stepNumber;
-
-    @Column(nullable = false)
-    private String stepContent;
-
-    @CreationTimestamp
-    private Timestamp createdAt;
-
-    @UpdateTimestamp
-    private Timestamp modifiedAt;
-}
